fix(starter-connector): align ConnectorProperties with Spring Boot
