+++ activiti-cloud-connectors/activiti-cloud-starter-connector/pom.xml
@@ -97,6 +97,11 @@
       <artifactId>spring-cloud-stream-binder-rabbit-test-support</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-configuration-processor</artifactId>
+      <optional>true</optional>
+    </dependency>
   </dependencies>
   <build>
     <plugins>
