+++ activiti-cloud-connectors/activiti-cloud-starter-connector/src/main/java/org/activiti/cloud/connectors/starter/channels/IntegrationResultDestinationBuilderImpl.java
@@ -3,12 +3,8 @@ package org.activiti.cloud.connectors.starter.channels;
 import org.activiti.cloud.api.process.model.IntegrationRequest;
 import org.activiti.cloud.connectors.starter.configuration.ConnectorProperties;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 
 public class IntegrationResultDestinationBuilderImpl implements IntegrationResultDestinationBuilder {
-
-    @Value("${ACT_INT_RES_CONSUMER:}")
-    private String resultDestinationOverride;
 
     private final ConnectorProperties connectorProperties;
 
@@ -19,17 +15,12 @@ public class IntegrationResultDestinationBuilderImpl implements IntegrationResul
 
     @Override
     public String buildDestination(IntegrationRequest event) {
+        String resultDestinationOverride = connectorProperties.getErrorDestinationOverride();
+
         String destination = (resultDestinationOverride == null || resultDestinationOverride.isEmpty())
                 ? "integrationResult" + connectorProperties.getMqDestinationSeparator() + event.getServiceFullName() : resultDestinationOverride;
 
         return destination;
     }
 
-    public String getResultDestinationOverride() {
-        return resultDestinationOverride;
-    }
-
-    public void setResultDestinationOverride(String resultDestinationOverride) {
-        this.resultDestinationOverride = resultDestinationOverride;
-    }
 }
