+++ activiti-cloud-connectors/activiti-cloud-starter-connector/src/main/resources/activiti-cloud-connector.properties
@@ -1,0 +1,8 @@
+activiti.cloud.connector.service-name=${spring.application.name}
+activiti.cloud.connector.service-type=${activiti.cloud.service.type:}
+activiti.cloud.connector.service-version=${activiti.cloud.service.version:}
+activiti.cloud.connector.app-name=${activiti.cloud.application.name:}
+activiti.cloud.connector.app-version=${activiti.cloud.application.version:}
+activiti.cloud.connector.mq-destination-separator=${activiti.cloud.mq.destination.separator:_}
+activiti.cloud.connector.result-destination-override=${ACT_INT_RES_CONSUMER:}
+activiti.cloud.connector.error-destination-override=${ACT_INT_ERR_CONSUMER:}
